Improve testing infrastructure
