+++ app/src/androidTest/java/ch/epfl/sweng/eventmanager/test/ticketing/MockTicketingModule.java
@@ -1,9 +1,8 @@
 package ch.epfl.sweng.eventmanager.test.ticketing;
 
 import ch.epfl.sweng.eventmanager.ticketing.TicketingServiceManager;
-import ch.epfl.sweng.eventmanager.ticketing.impl.TicketingServiceManagerImpl;
-import dagger.Binds;
 import dagger.Module;
+import dagger.Provides;
 
 import javax.inject.Singleton;
 
@@ -11,8 +10,16 @@ import javax.inject.Singleton;
  * @author Louis Vialar
  */
 @Module
-public abstract class MockTicketingModule {
-    @Binds
+public class MockTicketingModule {
+    @Provides
     @Singleton
-    abstract TicketingServiceManager providesTicketingServiceManager(MockTicketingServiceManager impl);
+    MockTicketingServiceManager providesMockTicketingServiceManager() {
+        return new MockTicketingServiceManager();
+    }
+
+    @Provides
+    @Singleton
+    TicketingServiceManager providesTicketingServiceManager(MockTicketingServiceManager impl) {
+        return impl;
+    }
 }
